admin plugin type isolation
