+++ packages/better-auth/src/plugins/admin/schema.ts
@@ -1,0 +1,39 @@
+import { type AuthPluginSchema } from "../../types";
+
+export const schema = {
+	user: {
+		fields: {
+			role: {
+				type: "string",
+				required: false,
+				input: false,
+			},
+			banned: {
+				type: "boolean",
+				defaultValue: false,
+				required: false,
+				input: false,
+			},
+			banReason: {
+				type: "string",
+				required: false,
+				input: false,
+			},
+			banExpires: {
+				type: "date",
+				required: false,
+				input: false,
+			},
+		},
+	},
+	session: {
+		fields: {
+			impersonatedBy: {
+				type: "string",
+				required: false,
+			},
+		},
+	},
+} satisfies AuthPluginSchema;
+
+export type AdminSchema = typeof schema;
