+++ packages/better-auth/src/plugins/admin/types.ts
@@ -1,0 +1,83 @@
+import { type Session, type User } from "../../types";
+import { type InferOptionSchema } from "../../types";
+import { type AccessControl, type Role } from "../access";
+import { type AdminSchema } from "./schema";
+
+export interface UserWithRole extends User {
+	role?: string;
+	banned?: boolean | null;
+	banReason?: string | null;
+	banExpires?: Date | null;
+}
+
+export interface SessionWithImpersonatedBy extends Session {
+	impersonatedBy?: string;
+}
+
+export interface AdminOptions {
+	/**
+	 * The default role for a user
+	 *
+	 * @default "user"
+	 */
+	defaultRole?: string;
+	/**
+	 * Roles that are considered admin roles.
+	 *
+	 * Any user role that isn't in this list, even if they have the permission,
+	 * will not be considered an admin.
+	 *
+	 * @default ["admin"]
+	 */
+	adminRoles?: string | string[];
+	/**
+	 * A default ban reason
+	 *
+	 * By default, no reason is provided
+	 */
+	defaultBanReason?: string;
+	/**
+	 * Number of seconds until the ban expires
+	 *
+	 * By default, the ban never expires
+	 */
+	defaultBanExpiresIn?: number;
+	/**
+	 * Duration of the impersonation session in seconds
+	 *
+	 * By default, the impersonation session lasts 1 hour
+	 */
+	impersonationSessionDuration?: number;
+	/**
+	 * Custom schema for the admin plugin
+	 */
+	schema?: InferOptionSchema<AdminSchema>;
+	/**
+	 * Configure the roles and permissions for the admin
+	 * plugin.
+	 */
+	ac?: AccessControl;
+	/**
+	 * Custom permissions for roles.
+	 */
+	roles?: {
+		[key in string]?: Role;
+	};
+	/**
+	 * List of user ids that should have admin access
+	 *
+	 * If this is set, the `adminRole` option is ignored
+	 */
+	adminUserIds?: string[];
+	/**
+	 * Message to show when a user is banned
+	 *
+	 * By default, the message is "You have been banned from this application"
+	 */
+	bannedUserMessage?: string;
+}
+
+export type InferAdminRolesFromOption<O extends AdminOptions | undefined> =
+	O extends { roles: Record<string, unknown> }
+		? keyof O["roles"]
+		: "user" | "admin";
